% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSIONFunctions.R
\name{addOption}
\alias{addOption}
\title{FUSION R command line interface -- Build text string for a command line option.}
\usage{
addOption(cl, opt, quote = FALSE)
}
\arguments{
\item{cl}{character string. Potentially contains other command line elements.}

\item{opt}{object containing the option value. The name of the object and the
value are used when constructing the option text.}

\item{quote}{boolean indicating the entire option should be enclosed in quotes.}
}
\value{
A (invisible) string containing the modified \code{cl} string.
}
\description{
This is a helper function used in the fusionwrapr package to build the text string
for command line options that require parameters. The resulting string is appended
to \code{cl}.
}
\examples{
\dontrun{
addOption(cl, zmin)
}
}
