% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSIONFunctions.R
\name{addRequired}
\alias{addRequired}
\title{FUSION R command line interface -- Build text string for a required parameter.}
\usage{
addRequired(cl, req, quote = FALSE)
}
\arguments{
\item{cl}{character string. Potentially contains other command line elements.}

\item{req}{object containing the value for the required parameter.}

\item{quote}{boolean indicating the entire parameter should be enclosed in quotes.}
}
\value{
A (invisible) string containing the modified \code{cl} string.
}
\description{
This is a helper function used in the fusionwrapr package to build the text string
for a required parameter. The resulting string is appended to \code{cl}.
}
\examples{
\dontrun{
addRequired(cl, "*.las")
}
}
