% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSIONFunctions.R
\name{buildCommand}
\alias{buildCommand}
\title{FUSION R command line interface -- Construct the command line.}
\usage{
buildCommand(cmd, options, required)
}
\arguments{
\item{cmd}{character string containing the program name and possibly the path.}

\item{options}{character string containing command line options.}

\item{required}{character string(s) containing required parameters for the command line.}
}
\value{
invisible string containing complete command line
}
\description{
This is a helper function used in the fusionwrapr package to construct the command line.
}
\examples{
\dontrun{
buildCommand("ClipData", "/minht:2.0", "*.las clip1.las")
}
}
