% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSIONFunctions.R
\name{checkRunSaveFile}
\alias{checkRunSaveFile}
\title{FUSION R command line interface -- Verify compatibility of function options.}
\usage{
checkRunSaveFile(runCmd, saveCmd, cmdFile)
}
\arguments{
\item{runCmd}{boolean indicating command line should be executed.}

\item{saveCmd}{boolean indicating command line should be written to a file.}

\item{cmdFile}{character string containing the name of the file to which commands
should be written.}
}
\value{
An (invisible) boolean TRUE, If the options are inconsistent, execution stops.
}
\description{
This is a helper function used in the fusionr package to verify the combination
of the runCmd, saveCmd, and cmdFile options for functions that build command line.
}
\examples{
\dontrun{
checkRunSaveFile(FALSE, TRUE, "test.bat")
}
}
