% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSIONFunctions.R
\name{ClipData}
\alias{ClipData}
\title{FUSION R command line interface -- Function to create command lines for the ClipData program.}
\usage{
ClipData(
  InputSpecifier = NULL,
  SampleFile = NULL,
  MinX = "-1000000000",
  MinY = "-1000000000",
  MaxX = "1000000000",
  MaxY = "1000000000",
  quiet = FALSE,
  verbose = FALSE,
  version = FALSE,
  newlog = FALSE,
  log = NULL,
  locale = FALSE,
  nolaszipdll = FALSE,
  shape = 0,
  decimate = NULL,
  ground = NULL,
  zmin = NULL,
  zmax = NULL,
  zpercent = NULL,
  height = FALSE,
  timemin = NULL,
  timemax = NULL,
  anglemin = NULL,
  anglemax = NULL,
  zero = FALSE,
  biaselev = NULL,
  return = NULL,
  class = NULL,
  ignoreoverlap = FALSE,
  line = NULL,
  noindex = FALSE,
  index = FALSE,
  lda = FALSE,
  nooffset = FALSE,
  cleanlas = FALSE,
  precision = NULL,
  use64bit = TRUE,
  runCmd = TRUE,
  saveCmd = TRUE,
  cmdFile = NULL,
  cmdClear = FALSE,
  echoCmd = FALSE
)
}
\arguments{
\item{InputSpecifier}{a}

\item{SampleFile}{a}

\item{MinX}{a}

\item{MinY}{a}

\item{MaxX}{a}

\item{MaxY}{a}

\item{quiet}{a}

\item{verbose}{a}

\item{version}{a}

\item{newlog}{a}

\item{log}{a}

\item{locale}{a}

\item{nolaszipdll}{a}

\item{shape}{a}

\item{decimate}{a}

\item{ground}{a}

\item{zmin}{a}

\item{zmax}{a}

\item{zpercent}{a}

\item{height}{a}

\item{timemin}{a}

\item{anglemin}{a}

\item{anglemax}{a}

\item{zero}{a}

\item{biaselev}{a}

\item{return}{a}

\item{class}{a}

\item{ignoreoverlap}{a}

\item{line}{a}

\item{noindex}{a}

\item{index}{a}

\item{lda}{a}

\item{nooffset}{a}

\item{cleanlas}{a}

\item{precision}{a}

\item{use64bit}{boolean value indicating 64-bit version of the program
specified in \code{name} should be used.}

\item{runCmd}{boolean indicating command line should be executed.}

\item{saveCmd}{boolean indicating command line should be written to a file.}

\item{cmdFile}{character string containing the name of the file to which commands
should be written.}

\item{cmdClear}{boolean indicating file for command should be deleted before the command
line is written.}

\item{echoCmd}{boolean indicating commadn line should be diasplayed.}

\item{timemmax}{a}
}
\value{
character string containing the \code{installPath} passed to the function.
}
\description{
/code{ClipData} creates command lines for the FUSION ClipData program and optionally executes them.
}
\examples{
\dontrun{
ClipData("*.las", "clip1.las", ground = "small.dtm", height = TRUE)
}
}
