% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSIONFunctions.R
\name{CloudMetrics}
\alias{CloudMetrics}
\title{FUSION R command line interface -- Function to create command lines for the CloudMetrics program.}
\usage{
CloudMetrics(
  InputSpecifier = NULL,
  OutputFile = NULL,
  quiet = FALSE,
  verbose = FALSE,
  version = FALSE,
  newlog = FALSE,
  log = NULL,
  locale = FALSE,
  nolaszipdll = FALSE,
  above = NULL,
  new = FALSE,
  firstinpulse = FALSE,
  firstreturn = FALSE,
  highpoint = FALSE,
  subset = FALSE,
  id = FALSE,
  rid = FALSE,
  pa = FALSE,
  minht = NULL,
  maxht = NULL,
  outlier = NULL,
  ignoreoverlap = FALSE,
  strata = NULL,
  intstrata = NULL,
  kde = NULL,
  rgb = NULL,
  use64bit = TRUE,
  runCmd = TRUE,
  saveCmd = TRUE,
  cmdFile = NULL,
  cmdClear = FALSE,
  echoCmd = FALSE
)
}
\arguments{
\item{InputSpecifier}{A}

\item{OutputFile}{A}

\item{quiet}{A}

\item{verbose}{A}

\item{version}{a}

\item{newlog}{a}

\item{log}{a}

\item{locale}{a}

\item{nolaszipdll}{a}

\item{above}{a}

\item{new}{a}

\item{firstinpulse}{a}

\item{firstreturn}{a}

\item{highpoint}{a}

\item{subset}{a}

\item{id}{a}

\item{rid}{a}

\item{pa}{a}

\item{minht}{a}

\item{maxht}{a}

\item{outlier}{a}

\item{ignoreoverlap}{a}

\item{strata}{a}

\item{intstrata}{a}

\item{kde}{a}

\item{rgb}{a}

\item{use64bit}{boolean value indicating 64-bit version of the program
specified in \code{name} should be used.}

\item{runCmd}{boolean indicating command line should be executed.}

\item{saveCmd}{boolean indicating command line should be written to a file.}

\item{cmdFile}{character string containing the name of the file to which commands
should be written.}

\item{cmdClear}{boolean indicating file for command should be deleted before the command
line is written.}

\item{echoCmd}{boolean indicating commadn line should be diasplayed.}
}
\value{
character string containing the \code{installPath} passed to the function.
}
\description{
/code{CloudMetrics} creates command lines for the FUSION CloudMetrics program and optionally executes them.
}
\examples{
\dontrun{
CloudMetrics("points/*.las", "test.csv", minht = 2.0)
}
}
