% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{echoCommand}
\alias{echoCommand}
\title{FUSION R command line interface -- Construct and echo the command line.}
\usage{
echoCommand(cmd, options, required, echoCmd)
}
\arguments{
\item{cmd}{character string containing the program name and possibly the path.}

\item{options}{character string containing command line options.}

\item{required}{character string(s) containing required parameters for the command line.}

\item{echoCmd}{boolean indicating command line should be echoed using message().}
}
\value{
boolean TRUE
}
\description{
This is a helper function used in the fusionwrapr package to construct and echo the command
line (depending of value of echoCmd parameter).
}
\examples{
\dontrun{
echoCommand("ClipData", "/minht:2.0", "*.las clip1.las", echoCmd = TRUE)
}
}
\keyword{internal}
