% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterData.R
\name{FilterData}
\alias{FilterData}
\title{FUSION R command line interface -- Applies various filters to LIDAR return data.}
\usage{
FilterData(
  filtertype = NULL,
  filterparams = NULL,
  windowsize = NULL,
  outputfile = NULL,
  datafile = NULL,
  quiet = FALSE,
  verbose = FALSE,
  version = FALSE,
  newlog = FALSE,
  log = NULL,
  locale = FALSE,
  nolaszipdll = FALSE,
  skipfilecheck = FALSE,
  lda = FALSE,
  layers = FALSE,
  index = FALSE,
  invert = FALSE,
  minsd = NULL,
  minpts = NULL,
  minrange = NULL,
  mingap = NULL,
  gapratio = NULL,
  class = NULL,
  ignoreoverlap = FALSE,
  precision = NULL,
  reclass = NULL,
  use64bit = TRUE,
  runCmd = TRUE,
  saveCmd = TRUE,
  cmdFile = NULL,
  cmdClear = FALSE,
  echoCmd = FALSE,
  comment = NULL
)
}
\arguments{
\item{filtertype}{character: Filtering algorithm used to remove returns from the DataFile(s)
outlier removes returns above or below the mean elevation plus or minus FilterParms * standard
deviation of the elevations
outlier2  More robust outlier removal...experimental
minimum   removes all returns except the return with the minimum elevation
maximum   removes all returns except the return with the maximum elevation}

\item{filterparams}{numeric: Parameters specific to the filtering method. For outlier this is the
multiplier applied to the standard deviation. For minimum and maximum, FilterParms is ignored
(but a value must be included...use 0).}

\item{windowsize}{numeric: Size of the window used to compute the standard deviation of elevations
or the minimum/maximum return.}

\item{outputfile}{character (\strong{required}): Name of the output file.}

\item{datafile}{character (\strong{required}): LIDAR data file template or the  name of a text file containing
a list of file names (must have .txt extension),}

\item{quiet}{boolean: Suppress all output during the run.}

\item{verbose}{boolean: Display all status information during the run.}

\item{version}{boolean: Report version information and exit with no processing.}

\item{newlog}{boolean: Erase the existing log file and start a new log}

\item{log}{character: Use the name specified for the log file.}

\item{locale}{boolean: Adjust program logic to input and output locale-specific numeric
formats (e.g. use a comma for the decimal separator).}

\item{nolaszipdll}{boolean: Suppress the use of the LASzip dll (c) Martin Isenburg...
removes support for compressed LAS (LAZ) files. This option
is only useful for programs that read or write point files.}

\item{skipfilecheck}{boolean: Skip logic that checks for valid point files and removes those that are invalid
prior to the start of point processing. This option is designed to overcome some limitations with windows
and wildcard processing. This option is only useful for programs that read point data.}

\item{lda}{boolean: Write sample files using FUSION's LDA format when using LAS input
files. The default behavior is to write data in LAS format when the input data are in LAS
format. When using input data in a format other than LAS, sample files are written in LDA format.}

\item{layers}{boolean: Output intermediate raster data layers.}

\item{index}{boolean: Create FUSION index files for the \code{outputfile}.}

\item{invert}{boolean: Inverts the elevations for points so the logic will work for points below ground. Use with
\code{filtertype="outlier2"}.}

\item{minsd}{numeric: Minimum standard deviation for points within a window for filtering to take place.
Default is 1.0 elevation units Used only with \code{filtertype="outlier"}.}

\item{minpts}{numeric: Minimum number of points in a window for filtering to take place. Can be used with
all filters. Must be at least 3 when used with  \code{filtertype="outlier"}. The default when used with \code{filtertype="outlier"}
is 3. The default for  \code{filtertype="minimum"} or  \code{filtertype="maximum"} is 1.}

\item{minrange}{numeric: Minimum range in elevations within a window for outlier filtering to take place.
Default is 150.0 elevation units. Used only with \code{filtertype="outlier2"}.}

\item{mingap}{numeric: Minimum vertical distance that define a gap. Used to isolate points above
the majority of points in the filter window. Used only with \code{filtertype="outlier2"}.}

\item{gapratio}{numeric: Proportion of points in window that can be above a vertical gap.
Ranges from 0.0 to 1.0. Used only with \code{filtertype="outlier2"}.}

\item{class}{character: "#,#,#,...": LAS files only: Specifies that only points with classification
values listed are to be included in the subsample to be considered as potential ground points.
Classification values should be separated by a comma. For example, class = "2,3,4,5" and can range from 0 to 31.
If the first character in string is ~, the list is interpreted as the classes you DO NOT want included
in the subsample. For example class = "~2,3" would include all class values EXCEPT 2 and 3.}

\item{ignoreoverlap}{boolean: Ignore points with the overlap flag set (LAS V1.4+ format).}

\item{precision}{character: "#,#,#": Control the scale factor used for X, Y, and Z
values in output LAS files. These values will override the values
in the source LAS files. There is rarely any need for the scale
parameters to be smaller than 0.001.}

\item{reclass}{numeric: Change the classification code for points identified as outliers and write them
to the output file. The value is the classification code assigned to the points. Only valid
when used with \code{filtertype="outlier"} or \code{filtertype="outlier2"}.}

\item{use64bit}{boolean: indicates 64-bit version of the program
should be used.}

\item{runCmd}{boolean: indicates command line should be executed.}

\item{saveCmd}{boolean: indicates command line should be written to a file.}

\item{cmdFile}{character: contains the name of the file to which commands
should be written.}

\item{cmdClear}{boolean: indicates file for commands should be deleted (cleared) before the command
line is written.}

\item{echoCmd}{boolean: indicates command line should be displayed.}

\item{comment}{character string containing comment to be written to command file before writing
the actual command. Only used when \code{runCmd = FALSE} and \code{saveCmd = TRUE}. When written,
there is always a blank line before the comment line in the command file.}
}
\value{
Return value depends on \code{runCmd}. if \code{runCmd = TRUE}, return value is
  the (invisible) integer value return from the operating system after running the command.
  if \code{runCmd = FALSE}, return value is the (invisible) command line.
}
\description{
\code{FilterData} creates command lines for the FUSION FilterData program and optionally executes them.
}
\section{Multiple Commands}{

  LTKFunctions generate a separate command for each item when the required parameters are lists.
  This is useful when you want to use the same command and options but different input data and
  output files. However, this works only when the switches and options are the same for all
  commands (not lists). If you try to use a list of switches and/or options and a list of input
  files, you will produce commands using all sets of switches and/or options for every input
  file.
}

\examples{
\dontrun{
FilterData("thinned_pts.las", 10, 100.0, "*.las")
}
}
\seealso{
Other LTKFunctions: 
\code{\link{CSV2Grid}()},
\code{\link{CanopyMaxima}()},
\code{\link{CanopyModel}()},
\code{\link{Catalog}()},
\code{\link{ClipDTM}()},
\code{\link{ClipData}()},
\code{\link{CloudMetrics}()},
\code{\link{Cover}()},
\code{\link{DTMDescribe}()},
\code{\link{DensityMetrics}()},
\code{\link{FirstLastReturn}()},
\code{\link{GridMetrics}()},
\code{\link{GridSample}()},
\code{\link{GridSurfaceCreate}()},
\code{\link{GridSurfaceStats}()},
\code{\link{GroundFilter}()},
\code{\link{IntensityImage}()},
\code{\link{JoinDB}()},
\code{\link{MergeDTM}()},
\code{\link{MergeData}()},
\code{\link{MergeRaster}()},
\code{\link{RepairGridDTM}()},
\code{\link{ReturnDensity}()},
\code{\link{SplitDTM}()},
\code{\link{SurfaceSample}()},
\code{\link{SurfaceStats}()},
\code{\link{TINSurfaceCreate}()},
\code{\link{ThinData}()},
\code{\link{TopoMetrics}()},
\code{\link{TreeSeg}()},
\code{\link{fusionwrapr}}
}
\concept{LTKFunctions}
