% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperFunctions.R
\name{GetSurfaceValues}
\alias{GetSurfaceValues}
\title{FUSION R command line interface -- Super-function to run SurfaceSample to get interpolated values from surface for (X,Y) locations.}
\usage{
GetSurfaceValues(
  df,
  xLabel = "X",
  yLabel = "Y",
  idLabel = "Value",
  surfaceFile = NULL
)
}
\arguments{
\item{df}{data frame containing columns for the identifier (optional), X, and Y.}

\item{xLabel}{character: Label for the column containing the X value.}

\item{yLabel}{character: Label for the column containing the Y value.}

\item{idLabel}{character: Label for the new column in \code{df} containing the surface value.}

\item{surfaceFile}{character: Name for the input surface files (PLANS DTM format). \code{surfacefile} may
be a wildcard or text list file (extension .txt).}
}
\value{
Returns a dataframe with an additional column containing the sampled surface values. If the
  \code{surfacefile} does not cover the location or contains invalid data, values for locations will be -1.0.
}
\description{
\code{GetSurfaceValues} builds an input file with the identifier (optional), X, and Y and runs SurfaceSample
to interpolate a value from the specified surface. Output is a CSV file with the value.
}
\section{Multiple Commands}{

  LTKFunctions generate a separate command for each item when the required parameters are lists.
  This is useful when you want to use the same command and options but different input data and
  output files. However, this works only when the switches and options are the same for all
  commands (not lists). If you try to use a list of switches and/or options and a list of input
  files, you will produce commands using all sets of switches and/or options for every input
  file.
}

\examples{
\dontrun{
GetSurfaceValues(df, "X", "Y", "ground.dtm")
}
}
