% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSIONFunctions.R
\name{isOpt}
\alias{isOpt}
\title{FUSION R command line interface -- test if object is NULL}
\usage{
isOpt(x)
}
\arguments{
\item{x}{object}
}
\value{
A (invisible) boolean value, TRUE indicates the object value is not NULL,
  FALSE indicates the object value is NULL.
}
\description{
This is a helper function used in the fusionr package to test the
value of an object: usually a function parameter.
}
