% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSIONFunctions.R
\name{programName}
\alias{programName}
\title{FUSION R command line interface -- Build text string for the program name.}
\usage{
programName(name, use64bit = FALSE)
}
\arguments{
\item{name}{character string with the base program name.}

\item{use64bit}{boolean value indicating 64-bit version of the program
specified in \code{name} should be used. If \code{setFUSIONpath} has
been called, the path is added to the return string.}
}
\value{
A (invisible) string containing the program name.
}
\description{
This is a helper function used in the fusionwrapr package to build the text string
for the program name. It uses the \code{use64bit} parameter to control use of
64-bit versions of command line programs and will include the FUSION install
path if it has been provided using \code{setFUSIONpath()}.
}
\examples{
\dontrun{
programName("ClipData", TRUE)
}
}
