% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSIONFunctions.R
\name{readLDA}
\alias{readLDA}
\title{FUSION R command line interface -- Read LDA format files and return data frame or LAS-class object}
\usage{
readLDA(fileName = NULL, type = "DF", LASTemplate = NULL)
}
\arguments{
\item{fileName}{character (\strong{required}): Name of the LDA file containing point data.}

\item{type}{character: Desired return type: "DF" for a data frame and "LAS" for LAS-class object.}

\item{LASTemplate}{character: File name (including path) for the LAS/LAZ file that will provide a template
for the LAS header. Used only when \code{type = "LAS"}.}
}
\value{
Return value depends on \code{type}. If \code{type = "DF"}, return value is
  a (invisible) data frame with the columns listed above. If \code{type = "LAS"}, return value is a
  (invisible) LAS-class object compatible with the \strong{lidR} package.
}
\description{
\code{readLDA} reads point data stored in FUSION's LDA format and returns either a data frame or
a LAS-class object. To create the LAS-class object, you must provide an existing LAS or LAZ
file to provide a template for the LAS header. Typically, this would be a file used to clip the
LDA file using \code{TreeSeg} (this is the only FUSION program that can't output LAS files). The
template file also provides coordinate system information for the LAS-class object.
}
\details{
When \code{type = "DF"} (default), the return is a data frame containing the following columns:
\enumerate{
  \item x
  \item y
  \item x
  \item pulse: pseudo-pulse number. This is generated using the GPS time when the LDA file is created
  \item return: return number
  \item angle: scan angle rank
  \item intensity: intensity
}

When \code{type = "LAS"}, the LAS-class object returned by \code{readLDA} is somewhat imperfect.
Many of the attributes for individual points are not populated with valid values because the LDA
format does not have all of the information required to fully populate the point records. In
particular, GPS time,number of returns in the pulse, and classification are not available.
NumberOfReturns for all point records is set to 1 for all point records. This may cause tools
designed to evaluate the validity of LAS data to throw warning since there may be points labeled
as return 2, 3, ... yet the NumberOfReturns will still be 1.
}
\examples{
\dontrun{
pts <- readLDA("points.lda")
las <- readLDA("points.lda", type = "LAS", LASTemplate = "source.las")
}
}
