% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSIONFunctions.R
\name{resetGlobalCommandOptions}
\alias{resetGlobalCommandOptions}
\title{FUSION R command line interface -- reset global options to control the dispatch of commands.}
\usage{
resetGlobalCommandOptions()
}
\value{
boolean TRUE
}
\description{
\code{resetGlobalCommandOptions} sets environment variables local to the \code{fusionwrapr} package that control
how individual commands are handled to their default values.
}
\details{
Default behavior for dispatching commands is use the 64-bit version of programs (if available) and run the
commands without echoing them to the console.
}
\examples{
resetGlobalCommandOptions()
}
