% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{runCommandFile}
\alias{runCommandFile}
\title{FUSION R command line interface -- Run a command file}
\usage{
runCommandFile(cmdFile = NULL, ...)
}
\arguments{
\item{cmdFile}{character: contains the name of the file to be run.}

\item{...}{additional parameter passed to system2().}
}
\value{
invisible boolean indicating success or failure for the initiation of the run.
}
\description{
\code{runCommandFile} runs a command file.
}
\examples{
\dontrun{
runCommandFile("Test/test.bat")
}
}
