% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{setFUSIONpath}
\alias{setFUSIONpath}
\title{FUSION R command line interface -- Set a package environment variable containing the FUSION install folder.}
\usage{
setFUSIONpath(installPath)
}
\arguments{
\item{installPath}{character string containing the FUSION install folder. A trailing forward
slash will be added if not included.}
}
\value{
character string containing the \code{installPath} passed to the function.
}
\description{
\code{setFUSIONpath} sets an environment variable local to the \code{fusionwrapr} package that specifies
the install folder for FUSION. This is necessary when the FUSION install folder has not been added
to the PATH environment variable.
}
\examples{
\dontrun{
setFUSIONpath("C:/FUSION/")
}
}
