% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{setGlobalCommandOptions}
\alias{setGlobalCommandOptions}
\title{FUSION R command line interface -- Set global options to control the dispatch of commands.}
\usage{
setGlobalCommandOptions(
  use64bit = NULL,
  runCmd = NULL,
  saveCmd = NULL,
  cmdFile = NULL,
  echoCmd = NULL
)
}
\arguments{
\item{use64bit}{boolean: indicates 64-bit version of the program should be used.}

\item{runCmd}{boolean: indicates command line should be executed.}

\item{saveCmd}{boolean: indicates command line should be written to a file. If this is \code{TRUE}, you
must also set \code{cmdFile} either in the global options or in individual command function calls.}

\item{cmdFile}{character: contains the name of the file to which commands should be written.}

\item{echoCmd}{boolean: indicates command line should be displayed.}
}
\value{
boolean TRUE
}
\description{
\code{setGlobalCommandOptions} sets environment variables local to the \code{fusionwrapr} package that control
how individual commands are handled. Use of \code{setGlobalCommandOptions} can simplify testing of commands and
the creation of batch files.
}
\details{
Default behavior for dispatching commands is use the 64-bit version of programs (if available) and run the
commands without echoing them to the console.
}
\examples{
setGlobalCommandOptions(runCmd = FALSE)
}
