% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSIONFunctions.R
\name{useLogFile}
\alias{useLogFile}
\title{FUSION R command line interface -- Write comments or other lines to command file}
\usage{
useLogFile(logFile = NULL, cmdFile = NULL, logClear = FALSE, cmdClear = FALSE)
}
\arguments{
\item{logFile}{character: Name of the log file.}

\item{cmdFile}{character: contains the name of the file to which commands
should be written.}

\item{logClear}{boolean: add a line to the command file to delete the \code{logFile} before
setting the environment variable for the log file.}

\item{cmdClear}{boolean: indicates file for command should be deleted before the command
line is written. Use this option with caution as it will wipe out the content of the
command file.}
}
\value{
invisible boolean indicating success or failure
}
\description{
\code{useLogFile} writes code to a command file (batch file) to set the log
file for FUSION tools. The normal behavior is to write output to a log file in the
FUSION install folder (LTKmaster.log).
}
\details{
Note that the \code{logFile} can contain spaces (the folder name, file name, or both).
However, using spaces is not recommended and may cause problems. If you include a folder
name, you can use the backslash or forwardslash characters as the path separator.

To reset logging behavior to use the log file in the FUSION install folder, use \code{useLogFile("")}.
}
\examples{
\dontrun{
useLogFile("NewLog.log", "test.bat", logClear = TRUE)
}
}
