% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSIONFunctions.R
\name{writeDTM}
\alias{writeDTM}
\title{FUSION R command line interface -- Write DTM format files from a matrix, RasterLayer, or SpatRaster object}
\usage{
writeDTM(
  x,
  fileName = NULL,
  description = "DTM written by fusionwrapr R package",
  xyunits = NULL,
  zunits = NULL,
  coordsys = NULL,
  zone = NULL,
  horizdatum = NULL,
  vertdatum = NULL,
  originX = NULL,
  originY = NULL,
  columnSpacing = NULL,
  rowSpacing = NULL,
  rotate = TRUE
)
}
\arguments{
\item{x}{(\strong{required}): data object containing the values to be written to the DTM file. This can be a simple
matrix, a SpatRaster object, or a SpatialLayer object. Values are assumed to be in rows with the first row and column
in the upper left corner. Use \code{rotate = FALSE} if values are in columns with the first column and row in the
lower left corner (same arrangement as FUSION's DTM files).}

\item{fileName}{character (\strong{required}): Name for the DTM format file.}

\item{description}{character: Descriptive name for the DTM. Default is "DTM written by fusionwrapr R package". Length
will be truncated to 60 characters.}

\item{xyunits}{character (\strong{required}): Units for LIDAR data XY (M for meters or F for feet).}

\item{zunits}{character (\strong{required}): Units for LIDAR data elevations (M for meters or F for feet).}

\item{coordsys}{numeric (\strong{required}): Coordinate system for LIDAR data:
0 for unknown
1 for UTM
2 for state plane)}

\item{zone}{numeric (\strong{required}): Coordinate system zone for LIDAR data (0 for unknown).}

\item{horizdatum}{numeric (\strong{required}): Horizontal datum:
0 for unknown
1 for NAD27
2 for NAD83}

\item{vertdatum}{numeric (\strong{required}): Vertical datum:
0 for unknown
1 for NGVD29
2 for NAVD88
3 for GRS80}

\item{originX}{numeric (\strong{required when \code{x} is matrix}): X value for origin (lower left point in \code{x}).}

\item{originY}{numeric (\strong{required when \code{x} is matrix}): Y value for origin (lower left point in \code{x}).}

\item{columnSpacing}{numeric (\strong{required when \code{x} is matrix}): Spacing between columns.}

\item{rowSpacing}{numeric (\strong{required when \code{x} is matrix}): Spacing between rows.}

\item{rotate}{boolean: Flag indicating grid of values needs to be rotated before being written to \code{fileName}. See the
description of \code{x} for details regarding the expected arrangement of values in the grid.}
}
\value{
Returns an invisible boolean value indicating success (TRUE) or failure (FALSE).
}
\description{
\code{writeDTM} writes FUSION's DTM format files from data contained in a matrix, RasterLayer, or SpatRaster object.
}
\examples{
\dontrun{
writeDTM(rast, "test.dtm")
}
}
\seealso{
Other helpers: 
\code{\link{readDTM}()},
\code{\link{readLDA}()}
}
\concept{helpers}
