% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSIONFunctions.R
\name{addSwitch}
\alias{addSwitch}
\title{FUSION R command line interface -- Build text string for a command line option.}
\usage{
addSwitch(cl, opt)
}
\arguments{
\item{cl}{character string. Potentially contains other command line elements.}

\item{opt}{object containing the switch value. The name of the object is used
when constructing the option text.}
}
\value{
A (invisible) string containing the modified \code{cl} string.
}
\description{
This is a helper function used in the fusionr package to build the text string
for command line switches. The resulting string is appended to \code{cl}.
}
\examples{
\dontrun{
addOption(cl, zmin)
}
}
