% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSIONFunctions.R
\name{ClipPlot}
\alias{ClipPlot}
\title{FUSION R command line interface -- Super-function to clip data for plots using the ClipData program.}
\usage{
ClipPlot(
  inputspecifier = NULL,
  samplefile = NULL,
  x = NULL,
  y = NULL,
  radius = NULL,
  shape = 1,
  ...
)
}
\arguments{
\item{inputspecifier}{character: LIDAR data file template, name of a text file containing
a list of file names (must have .txt extension), or a
FUSION catalog file.}

\item{samplefile}{character: Name for plot clip file (extension will be added).
\code{samplefile} cannot contain spaces. To save compressed LAS files, specify the .laz extension.}

\item{x}{numeric: Easting value for plot location.}

\item{y}{numeric: Northing value for plot location.}

\item{radius}{numeric: Radius for round (circular) plots or half width for square plots.}

\item{shape}{numeric: default = 1: Shape of the sample area (0 = rectangle, 1 = circle).}

\item{...}{Additional parameters that will be passed to \code{ClipData}.}
}
\value{
Return value depends on \code{runCmd}. if \code{runCmd = TRUE}, return value is
  the (invisible) integer value return from the operating system after running the command.
  if \code{runCmd = FALSE}, return value is the (invisible) command line.
}
\description{
\code{ClipPlot} creates command lines for the FUSION ClipData program and optionally executes them.
Command lines are designed to clip a circular or square area centered in the \code{(x,y)}.
}
\examples{
\dontrun{
ClipPlot("*.las", "plot_0001.las", 435895.9, 5669341.4, 0, ground = "small.dtm", height = TRUE)
}
}
