% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSIONFunctions.R
\name{verifyFolder}
\alias{verifyFolder}
\title{FUSION R command line interface -- Verify existence of a folder}
\usage{
verifyFolder(folder)
}
\arguments{
\item{folder}{folder name}
}
\value{
An (invisible) boolean TRUE, If the folder does not exist and cannot
  be created, execution stops.
}
\description{
This is a helper function used in the fusionr package to verify
the existence of a folder and create the folder if it does not exist.
verifyFolder is normally called with a folder extracted using
\code{dirname()} when an output file is needed for a function.
}
\examples{
\dontrun{
verifyFolder("/test")
}
}
